requirements.txt
openai-whisper>=20231117
moviepy>=1.0.3
numpy>=1.21.0
Pillow>=8.3.0
ffmpeg-python>=0.2.0